Distance calculation now fixed.

+++ Assignment1/FeatureVectorDataset.py
@@ -40,4 +40,4 @@ class FeatureVectorDataset:
 		Sample IDs start with 0 and are consecutive.
 		The channel order of samples must be RGB.
 		Throws an error if the sample does not exist."""
-		return self.data[sid, :], self.labels[sid]
+		return self.data[sid, :].astype('int32'), self.labels[sid]
